[NVWS] Move partition metadata readers to IR (#11181)
commit that didn't get into
https://github.com/triton-lang/triton/pull/11116

patch layering per
https://github.com/triton-lang/triton/pull/11116#discussion_r3717710147
