+++ lib/Dialect/TritonGPU/Transforms/WarpSpecialization/PartitionAttrs.h
@@ -1,8 +1,7 @@
 #ifndef TRITON_LIB_DIALECT_TRITONGPU_TRANSFORMS_WARPSPECIALIZATION_PARTITIONATTRS_H_
 #define TRITON_LIB_DIALECT_TRITONGPU_TRANSFORMS_WARPSPECIALIZATION_PARTITIONATTRS_H_
 
-#include "mlir/Support/LLVM.h"
-#include "llvm/ADT/SetVector.h"
+#include "triton/Dialect/TritonGPU/IR/Partitioning.h"
 #include <optional>
 
 namespace mlir {
@@ -16,15 +15,10 @@ class ForOp;
 
 namespace mlir::triton::gpu {
 
-inline constexpr char kPartitionAttrName[] = "ttg.partition";
-inline constexpr char kPartitionOutputsAttrName[] = "ttg.partition.outputs";
 inline constexpr char kPartitionStagesAttrName[] = "ttg.partition.stages";
 inline constexpr char kWarpSpecializeTagAttrName[] = "ttg.warp_specialize.tag";
 
-SetVector<int> getPartitionIds(Operation *op);
-SmallVector<SetVector<int>, 4> getPartitionOutputs(Operation *op);
 SetVector<int> getPartitionIds(OpOperand *use);
-bool hasPartition(Operation *op);
 bool hasWarpSpecializeTag(Operation *op);
 std::optional<int> getWarpSpecializeTag(Operation *op);
